fix: get correct transport when running on default port
This fixes capturing manual browsers on Cloud9.

Socket.io always uses port numbers in the active sockets map (even for default ports), so we need to append the port.

I also changed to always use "http", as Karma does not support https and Cloud9 does (it uses a proxy), so the main frame is https, but the socket.io connection runs on unsecured http.