fix(adapter): Remove incorrect function and its call.
The string tested in the function isTopLevelSuite() is not used in jasmine. The
correct string would be Jasmine__TopLevel__Suite (two underscores). So the function
always returns false, and we negate it at the call site, to the child is always
added. That's the correct behavior because we already have a root node matching
the Jasmine top level suite. The test and the function is unnecessary.