fix(adapter): don't duplicate multiline failure messages
The current code to handle removing error messages for formatting does not
properly handle multiline error messages and results in the message being
repeated at the top of the stack trace. The code only searched for the message
in the first stack line, which meant it would never be found because the message
was multiple lines, which meant that the message would be pre-pended by the
message not in stack logic.

The existing multiline test actually triggers this behavior, but because it used
toMatch it only verified that the message was included, not that it only existed
once.